requirements.txt
Telethon==1.22.0
requests==2.25.1
GitPython==3.1.18
